% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitman_replication.R
\name{hitman_replication}
\alias{hitman_replication}
\title{High-throughput approach for testing directional replication}
\usage{
hitman_replication(
  tab,
  cols = 1:4,
  reorder.rows = FALSE,
  fdr.method = c("BH", "BY"),
  prefix = NA
)
}
\arguments{
\item{tab}{Matrix-like object with statistical and p-value columns. Only the signs of the statistics columns are used.
The object should not have missing values.}

\item{cols}{Vector of column indices or names in the order of \code{c(stat1, p1, stat2, p2)}.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{fdr.method}{Character string; either "BH" for Benjamini-Hochberg or "BY" for Benjamini-Yekutieli.
See stats::p.adjust.}

\item{prefix}{Character string of length one with prefix of returned columns, e.g. if \code{prefix="repl"}, returned columns will be \code{c("repl.chisq", "repl.p")}.
Prefix is not added if it is \code{NA}.}
}
\value{
Data frame whose rows correspond to the rows of \code{tab} with the same row names and whose columns are
\describe{
\item{chisq}{Chi-square for replication on 1 degreee of freedom.}
\item{p}{P-value for replication}
\item{FDR}{FDR for replication}
}
}
\description{
High-throughput approach for testing replication of two base studies, where each study applies a two-sided test whose
signed statistic and p-value are supplied via \code{tab}, and its desired to find rows where there is replication in a common direction.
}
\details{
If \code{tab}'s row names are \code{NULL}, then \code{reorder.rows} must be \code{FALSE}.

Larger chi-square values are more significant.
}
