% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_signif_mediation.R
\name{joint_signif_mediation}
\alias{joint_signif_mediation}
\title{Joint significant mediation test}
\usage{
joint_signif_mediation(E, M, Y, covariates = NULL)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{Numeric vector of mediators with one element per sample.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes. Only continuous, normally distributed outcomes
currently supported.}

\item{covariates}{Numeric vector with one element per sample or matrix-like object with rows corresponding
to samples and columns to covariates to be adjusted for.}
}
\value{
Data frame with columns \code{EMY.p}, \code{EM.p}, and \code{MY.p}.
}
\description{
Joint significant mediation test.
}
\details{
\code{E} and \code{Y} cannot have \code{NA}s. \code{M} may have some \code{NA}s, but rows that have
less non-missing values than \code{5 + ncol(covariates)} will be filtered out, and if \code{verbose=TRUE},
a message will be written with the number of rows filtered out.

Larger chi-square values are more significant.
}
\seealso{
\url{https://github.com/cedricbatailler/JSmediation}.
}
