% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotman.R
\name{lotman}
\alias{lotman}
\title{Low-throughput mediation analysis (Lotman)}
\usage{
lotman(E, M, Y, covariates = NULL, exponent = 1.24,
  check.names = TRUE)
}
\arguments{
\item{E}{A numeric vector of exposures.}

\item{M}{A numeric matrix-like data object with one row per feature and one column per sample of mediators.}

\item{Y}{A numeric vector of \code{length(E)} of outcomes. Only continuous, normally distributed outcomes
currently supported.}

\item{covariates}{Numeric vector with one element per sample or matrix-like object with rows corresponding
to samples and columns to covariates to be adjusted for.}

\item{exponent}{Numeric exponent for \code{max(p1, p2)}. Must be \code{0<=exponent<=2}.
You should understand this if you change it.}

\item{check.names}{Logical; should \code{names(E)==colnames(M) & colnames(M)==names(Y)} be checked?}
}
\value{
Data frame with columns
\describe{
\item{EMY.p}{Overall p-value for mediation}
\item{EMY.FDR}{Overall FDR for mediation}
\item{EM_dir.p}{p-value for E-->M accounting for direction of mediation}
\item{MY_dir.p}{p-value for M-->Y accounting for direction of mediation}
\item{EM.z}{z-score for E-->M, not accounting for direction}
\item{EM.p}{p-value for E-->M, not accounting for direction}
\item{MY.z}{z-score for M-->Y, not accounting for direction}
\item{MY.p}{p-value for M-->Y, not accounting for direction}
}
}
\description{
Low-throughput mediation analysis to test if rows of \code{M} mediate the effect of exposure \code{E} on outcome
\code{Y}.
}
\details{
\code{E} and \code{Y} cannot have \code{NA}s.
}
