% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_barfield.R
\name{sim_barfield}
\alias{sim_barfield}
\title{Simulate & test mediation methods as in Barfield et al. (2017)}
\usage{
sim_barfield(med.fnm, b1t2.v = c(0, 0.14, 0.39), t1 = 0.59,
  alpha = 0.05, nsamp = 50, nsim = 10^4, ngene = 0, seed = 0,
  verbose = TRUE, ...)
}
\arguments{
\item{med.fnm}{Quoted name of mediation function to test. The function must accept parameters and output matrix
with mediation p-value in column \code{"EMY.p"}, like \code{\link{hitman}}.}

\item{b1t2.v}{Numeric vector of values that both theta2 (\code{"t2"}) and beta1 (\code{"b1"}) take.}

\item{t1}{Numeric value of theta2 i.e. the effect of the exposure on the outcome.}

\item{alpha}{Alpha level.}

\item{nsamp}{Number of samples.}

\item{nsim}{Number of simulations.}

\item{ngene}{Number of genes other than that of primary interest to simulate.}

\item{seed}{Random seed for reprodubility.}

\item{verbose}{Logical; should the number of simulations be printed every 100 simulations?}

\item{...}{Sent to \code{med.fcn}}
}
\value{
Matrix with proportion of significant calls for every combination of \code{t2} and \code{b1}.
}
\description{
Simulate & test mediation methods as in Barfield et al. (2017). If multiple genes are simulated, only 1st
is tested. For each simulation, count proportion of p-values < \code{alpha}, and at end calculate mean proportion.
}
\references{
Barfield R, Shen J, Just AC, Vokonas PS, Schwartz J, Baccarelli AA, VanderWeele TJ, Lin X.
Testing for the indirect effect under the null for genome-wide mediation analyses. Genet Epidemiol.
2017 Dec;41(8):824-833.
}
