% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_min.R
\name{screen_min}
\alias{screen_min}
\title{ScreenMin multiple hypthesis testing for union hypothesis tests}
\usage{
screen_min(p.tab, thresh = 0.05, reorder.rows = TRUE, prefix = NULL,
  adjust.method = "BH")
}
\arguments{
\item{p.tab}{Numberic matrix with two columns of p-values.
Each row is an analyte. No \code{NA}'s allowed.}

\item{thresh}{Numeric threshold, termed \code{c} in Djordjilovic et al. (2019). Min p-values > \code{thresh}
are assigned adjusted p-values of one.}

\item{reorder.rows}{Logical, should rows be reordered by p-value?}

\item{prefix}{Character string to add to beginning of column names. \code{NULL} does not add a prefix.}

\item{adjust.method}{Method used to adjust the p-values for multiple testing. Options, in increasing conservatism, 
include \code{"none"}, \code{"BH"}, \code{"BY"}, and \code{"holm"}. See \code{\link[stats]{p.adjust}} for the complete
list of options. A \code{NULL} value will result in the default adjustment method, which is \code{"BH"}.}
}
\description{
ScreenMin multiple hypthesis testing for union hypothesis tests from Djordjilovic et al. (2018).
}
\references{
Djordjilovic V, Page CM, Gran JM, Nost TH, Sandanger TM, Veierød MB, Thoresen M.
Global test for high-dimensional mediation: Testing groups of potential mediators. Stat Med. 2019 Aug 15;38(18):3346-3360.
}
