% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_omics.R
\name{sim_omics}
\alias{sim_omics}
\title{Simulate & test mediation methods on omics dataset}
\usage{
sim_omics(
  b1 = 1,
  t2 = b1,
  t1 = 5,
  nsamp = 15,
  ngene = 100,
  FDR = 0.25,
  Sigma = diag(ngene),
  prop.consistent = 1/ngene,
  prop.inconsistent = 0,
  prop.1c = 0,
  sd.mn = 0.5,
  nsim = 10^3,
  fdr.method = c("BH", "BY"),
  seed = 0,
  verbose = TRUE
)
}
\arguments{
\item{b1}{Numeric value that beta1 (\code{"b1"}), the EM coefficient, takes for non-null mediators.}

\item{t2}{Numeric value that theta2 (\code{"t2"}), the MY coefficient, takes for non-null mediators.}

\item{t1}{Numeric value of theta1 i.e. the effect of the exposure on the outcome.}

\item{nsamp}{Number of samples.}

\item{ngene}{Number of genes.}

\item{FDR}{FDR threshold to apply.}

\item{Sigma}{Gene covariance matrix.}

\item{prop.consistent}{Proportion of genes that are consistent mediators. Must be at least \code{1/ngene}.}

\item{prop.inconsistent}{Proportion of genes that are inconsistent mediators.}

\item{prop.1c}{Proportion of genes where exactly one component of the test (E --> M or M --> Y given E) holds.}

\item{sd.mn}{Numeric standard deviation of measurement noise.}

\item{nsim}{Number of simulations.}

\item{fdr.method}{Character string; either "BH" for Benjamini-Hochberg or "BY" for Benjamini-Yekutieli.
See stats::p.adjust.}

\item{seed}{Random seed for reprodubility.}

\item{verbose}{Logical; should the number of simulations be printed every 100 simulations?}
}
\value{
Matrix with proportion of significant calls for every method for true and null mediators.
}
\description{
Simulate & test mediation methods on omics dataset.
}
